Reduced constant memory

+++ src/monopole_interactions/p2p_cuda_kernel.hpp
@@ -6,7 +6,7 @@
 namespace octotiger {
 namespace fmm {
     namespace monopole_interactions {
-        __host__ void copy_stencil_to_p2p_constant_memory(const double *stencil_masks, const size_t full_stencil_size);
+        __host__ void copy_stencil_to_p2p_constant_memory(const float *stencil_masks, const size_t full_stencil_size);
         __host__  void copy_constants_to_p2p_constant_memory(const double *constants, const size_t constants_size);
         __global__ void cuda_p2p_interactions_kernel(
             const double (&local_monopoles)[NUMBER_LOCAL_MONOPOLE_VALUES],
